feat(relationships): new stable api
