+++ packages/beta/src/types.ts
@@ -3,137 +3,13 @@
 import {
   CogniteExternalId,
   CogniteInternalId,
-  CreatedAndLastUpdatedTimeFilter,
-  DateRange,
   FilterQuery,
   IdEither,
-  Label,
-  Timestamp,
-  Range,
-  LabelFilter,
   SinglePatchString,
-  CreatedAndLastUpdatedTime,
 } from '@cognite/sdk';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible
-
-export type RelationshipResourceType =
-  | 'asset'
-  | 'timeSeries'
-  | 'file'
-  | 'event'
-  | 'sequence';
-
-export interface ExternalRelationship {
-  /**
-   * External id of the relationship, must be unique within the project
-   */
-  externalId: CogniteExternalId;
-  /**
-   * External id of the CDF resource that constitutes the relationship source
-   */
-  sourceExternalId: CogniteExternalId;
-  /**
-   * Enum: "asset" "timeSeries" "file" "event" "sequence".
-   * The CDF resource type of the relationship source. Must be one of the specified values.
-   */
-  sourceType: RelationshipResourceType;
-  /**
-   * External id of the CDF resource that constitutes the relationship target.
-   */
-  targetExternalId: CogniteExternalId;
-  /**
-   * Enum: "asset" "timeSeries" "file" "event" "sequence".
-   * The CDF resource type of the relationship target. Must be one of the specified values.
-   */
-  targetType: RelationshipResourceType;
-  /**
-   * Time when the relationship became active.
-   * If there is no startTime, relationship is active from the beginning of time until endTime.
-   */
-  startTime?: Timestamp;
-  /**
-   * Time when the relationship became inactive. If there is no endTime,
-   * relationship is active from startTime until the present or any point in the future.
-   * If endTime and startTime are set, then endTime must be strictly greater than startTime
-   */
-  endTime?: Timestamp;
-  /**
-   * Confidence value of the existence of this relationship.
-   * Generated relationships should provide a realistic score on the likelihood of the existence of the relationship.
-   * Relationships without a confidence value can be interpreted at the discretion of each project.
-   */
-  confidence?: number;
-  /**
-   * The id of the dataset this relationship belongs to
-   */
-  dataSetId?: CogniteInternalId;
-  /**
-   * A list of the labels associated with this resource item
-   */
-  labels?: Label[];
-}
-
-export interface Relationship
-  extends ExternalRelationship,
-    CreatedAndLastUpdatedTime {}
-
-export interface RelationshipsFilterRequest extends FilterQuery {
-  /**
-   * Filter on relationships with exact match.
-   * Multiple filter elements in one property, e.g. sourceExternalIds: [ "a", "b" ],
-   * will return all relationships where the sourceExternalId field is either a or b.
-   * Filters in multiple properties will return the relationships that match all criteria.
-   * If the filter is not specified it default to an empty filter.
-   */
-  filter?: RelationshipsFilter;
-}
-
-export interface RelationshipsFilter extends CreatedAndLastUpdatedTimeFilter {
-  /**
-   * Include relationships that have any of these values in their sourceExternalId field
-   */
-  sourceExternalIds?: CogniteExternalId[];
-  /**
-   * Items Enum: "asset" "timeSeries" "file" "event" "sequence".
-   * Include relationships that have any of these values in their sourceType field
-   */
-  sourceTypes?: RelationshipResourceType[];
-  /**
-   * Include relationships that have any of these values in their targetExternalId field
-   */
-  targetExternalIds?: CogniteExternalId[];
-  /**
-   * Items Enum: "asset" "timeSeries" "file" "event" "sequence".
-   * Include relationships that have any of these values in their targetType field
-   */
-  targetTypes?: RelationshipResourceType[];
-  /**
-   * Only include relationships that reference these specific dataSet IDs
-   */
-  dataSetIds?: IdEither[];
-  startTime?: DateRange;
-  endTime?: DateRange;
-  /**
-   * Confidence range to filter for
-   */
-  confidence?: Range<number>;
-  /**
-   * Limits results to those active at any point within the given time range, i.e.
-   * if there is any overlap in the intervals [activeAtTime.min, activeAtTime.max] and [startTime, endTime],
-   * where both intervals are inclusive. If a relationship does not have a startTime,
-   * it is regarded as active from the beginning of time by this filter.
-   * If it does not have an endTime is will be regarded as active until the end of time.
-   * Similarly, if a min is not supplied to the filter, the min will be implicitly set to the beginning of time,
-   * and if a max is not supplied, the max will be implicitly set to the end of time.
-   */
-  activeAtTime?: DateRange;
-  /**
-   * Return only the resource matching the specified label constraints
-   */
-  labels?: LabelFilter;
-}
 
 export type ContextJobStatus = 'QUEUED' | 'RUNNING' | 'COMPLETED' | 'FAILED';
 
@@ -488,3 +364,5 @@ export interface EntityMatchingFilter {
 export interface EntityMatchingFilterRequest extends FilterQuery {
   filter?: EntityMatchingFilter;
 }
+// This file is here mostly to allow apis to import { ... } from '../../types';
+// Overriding types should probably be done in their respective API endpoint files, where possible
